% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tikzsetup.R
\name{tikzmagic}
\alias{tikzmagic}
\title{Lanch this function after installation of the tikzsetup.
This should be done in clear R, when all packages are unloaded.}
\usage{
tikzmagic()
}
\value{
nothing
}
\description{
Lanch this function after installation of the tikzsetup.
This should be done in clear R, when all packages are unloaded.
}
\examples{
tikzmagic()
}

