% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{tikzsetup}
\alias{tikzsetup}
\alias{tikzsetup-package}
\title{tikzsetup}
\usage{
tikzsetup(compiler = "pdftex", lang = "russian",
  doc_class_options = "12pt,a4paper", message = FALSE, warning = FALSE,
  setJustPlot = TRUE, pt_size = 12)
}
\arguments{
\item{compiler}{currently pdftex and xelatex are supported}

\item{lang}{the language}

\item{doc_class_options}{additional latex document class options}

\item{message}{logical, should messages be activated, FALSE by default}

\item{warning}{logical, should warnings be activated, FALSE by default}

\item{setJustPlot}{logical, TRUE by default. Sometimes tikzsetup()
is called from a big-big document with a bunch of packages which may
interfere with tikzsetup(). In this one may use something like
\\ifdef{\JustPlot}{}{some dangerous command}}
}
\value{
nothing
}
\description{
tikzsetup

Setup tikz device for russian Rmd/Rnw files
}
\details{
Setup tikz device for russian Rmd/Rnw files
}
\examples{
tikzsetup()
tikzsetup(lang="polish")
}
\author{
Boris Demeshev
}

